Port TripFetcher to coroutines and consolidate one-shot fetches
Unify the trip data layer's two network sources on one async idiom and
document their split: Pollers are the time-driven half (recurring,
lifecycle-bound re-fetch of volatile vehicle status), TripFetcher the
demand-driven half (one-shot memoization of immutable schedules/shapes
plus explicit refreshes).

- Replace TripFetcher's thread pool + Handler with MainScope +
  withContext(Dispatchers.IO), preserving the old 2-thread rate limit
  explicitly via a Semaphore so route-poll backfills can't fan out into
  dozens of simultaneous API requests.
- Move fetchTripDetailsOnce from Pollers.kt to TripFetcher — it's
  demand-driven, not a poller; drop the now-unneeded @file:JvmName.
- Add doc headers on both files stating the volatile-vs-immutable
  contract.
